enh(UI): Integrate the new Form component (#11318)
* Move Web SSO form

* Move OpenID form

* Fix eslint errors

* Migrate Password security pollicy to use the Form builder

* Fix eslint errors

* Update dependency
