+++ www/front_src/src/Authentication/Openid/Form/index.tsx
@@ -1,10 +1,8 @@
-import { Formik, FormikValues } from 'formik';
+import { FormikValues } from 'formik';
 import { useTranslation } from 'react-i18next';
 import { isEmpty, isNil, pick, pipe, values, or, all, not } from 'ramda';
 
-import { makeStyles } from '@mui/styles';
-
-import { useRequest, useSnackbar } from '@centreon/ui';
+import { useRequest, useSnackbar, Form } from '@centreon/ui';
 
 import useValidationSchema from '../useValidationSchema';
 import {
@@ -15,31 +13,25 @@ import {
 import { putProviderConfiguration } from '../../api';
 import { OpenidConfiguration, OpenidConfigurationToAPI } from '../models';
 import FormButtons from '../../FormButtons';
-import Inputs from '../../FormInputs';
-import { categories } from '../..';
+import { groups } from '../..';
 import { Provider } from '../../models';
 import { adaptOpenidConfigurationToAPI } from '../../api/adapters';
 
 import { inputs } from './inputs';
 
-const useStyles = makeStyles((theme) => ({
-  formContainer: {
-    margin: theme.spacing(2, 0, 0),
-  },
-}));
-
 interface Props {
   initialValues: OpenidConfiguration;
+  isLoading: boolean;
   loadOpenidConfiguration: () => void;
 }
 
 const isNilOrEmpty = (value): boolean => or(isNil(value), isEmpty(value));
 
-const Form = ({
+const OpenidForm = ({
   initialValues,
   loadOpenidConfiguration,
+  isLoading,
 }: Props): JSX.Element => {
-  const classes = useStyles();
   const { t } = useTranslation();
 
   const { sendRequest } = useRequest({
@@ -89,21 +81,17 @@ const Form = ({
   };
 
   return (
-    <Formik
-      enableReinitialize
-      validateOnBlur
-      validateOnMount
+    <Form<OpenidConfiguration>
+      Buttons={FormButtons}
+      groups={groups}
       initialValues={initialValues}
+      inputs={inputs}
+      isLoading={isLoading}
+      submit={submit}
       validate={validate}
       validationSchema={validationSchema}
-      onSubmit={submit}
-    >
-      <div className={classes.formContainer}>
-        <Inputs categories={categories} inputs={inputs} />
-        <FormButtons />
-      </div>
-    </Formik>
+    />
   );
 };
 
-export default Form;
+export default OpenidForm;
