+++ www/front_src/src/Authentication/WebSSO/Form/index.tsx
@@ -1,9 +1,6 @@
-import { Formik } from 'formik';
 import { useTranslation } from 'react-i18next';
 
-import { makeStyles } from '@mui/styles';
-
-import { useRequest, useSnackbar } from '@centreon/ui';
+import { useRequest, useSnackbar, Form } from '@centreon/ui';
 
 import useValidationSchema from '../useValidationSchema';
 import {
@@ -12,30 +9,24 @@ import {
 } from '../translatedLabels';
 import { putProviderConfiguration } from '../../api';
 import { WebSSOConfiguration, WebSSOConfigurationToAPI } from '../models';
-import FormButtons from '../../FormButtons';
-import Inputs from '../../FormInputs';
-import { categories } from '../..';
+import { groups } from '../..';
 import { Provider } from '../../models';
 import { adaptWebSSOConfigurationToAPI } from '../../api/adapters';
+import FormButtons from '../../FormButtons';
 
 import { inputs } from './inputs';
 
-const useStyles = makeStyles((theme) => ({
-  formContainer: {
-    margin: theme.spacing(2, 0, 0),
-  },
-}));
-
 interface Props {
   initialValues: WebSSOConfiguration;
+  isLoading: boolean;
   loadWebSSOonfiguration: () => void;
 }
 
-const Form = ({
+const WebSSOForm = ({
   initialValues,
   loadWebSSOonfiguration,
+  isLoading,
 }: Props): JSX.Element => {
-  const classes = useStyles();
   const { t } = useTranslation();
 
   const { sendRequest } = useRequest({
@@ -64,20 +55,16 @@ const Form = ({
       .finally(() => setSubmitting(false));
 
   return (
-    <Formik
-      enableReinitialize
-      validateOnBlur
-      validateOnMount
+    <Form<WebSSOConfiguration>
+      Buttons={FormButtons}
+      groups={groups}
       initialValues={initialValues}
+      inputs={inputs}
+      isLoading={isLoading}
+      submit={submit}
       validationSchema={validationSchema}
-      onSubmit={submit}
-    >
-      <div className={classes.formContainer}>
-        <Inputs categories={categories} inputs={inputs} />
-        <FormButtons />
-      </div>
-    </Formik>
+    />
   );
 };
 
-export default Form;
+export default WebSSOForm;
